requirements.txt
streamlit>=1.24.0
mistralai>=0.0.7
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
aiohttp>=3.8.0